Move EXIF processing into metadata module
