+++ mobile/apps/photos/lib/module/metadata/exif.dart
