+++ mobile/apps/photos/lib/module/upload/upload_metadata.dart
@@ -8,8 +8,8 @@ import "package:photos/gateways/collections/models/metadata.dart";
 import "package:photos/models/file/file.dart";
 import "package:photos/models/file/file_type.dart";
 import "package:photos/models/metadata/file_magic.dart";
+import "package:photos/module/metadata/exif.dart";
 import "package:photos/module/upload/model/media_upload_data.dart";
-import "package:photos/utils/exif_util.dart";
 import "package:photos/utils/panorama_util.dart";
 
 Future<Map<String, dynamic>> buildUploadMetadata(
