+++ mobile/apps/photos/lib/services/sync/offline_import_metadata_service.dart
@@ -9,11 +9,11 @@ import "package:photos/events/files_updated_event.dart";
 import "package:photos/events/local_photos_updated_event.dart";
 import "package:photos/models/file/file.dart";
 import "package:photos/models/location/location.dart";
+import "package:photos/module/metadata/exif.dart";
 import "package:photos/module/upload/model/media_upload_data.dart";
 import "package:photos/module/upload/upload_metadata.dart";
 import "package:photos/service_locator.dart";
 import "package:photos/utils/embedded_media_location.dart";
-import "package:photos/utils/exif_util.dart";
 import "package:photos/utils/file_util.dart";
 
 class OfflineImportMetadataService {
